moved context wrapper in bpemb embedding model
