+++ deepparse/parser/tools.py
@@ -1,13 +1,9 @@
-import contextlib
 import math
 import os
-import warnings
 from typing import List, OrderedDict, Tuple
 
 import numpy as np
-import requests
 import torch
-from urllib3.exceptions import InsecureRequestWarning
 
 
 def validate_if_new_prediction_tags(checkpoint_weights: dict) -> bool:
@@ -142,41 +138,3 @@ def infer_model_type(checkpoint_weights: OrderedDict, attention_mechanism: bool)
         attention_mechanism = True
 
     return model_type, attention_mechanism
-
-
-@contextlib.contextmanager
-def no_ssl_verification():
-    """Context Manager to disable SSL verification when using ``requests`` library.
-
-    Reference: https://gist.github.com/ChenTanyi/0c47652bd916b61dc196968bca7dad1d.
-
-    Can be removed after https://github.com/bheinzerling/bpemb/issues/63 is resolved.
-    """
-    opened_adapters = set()
-    old_merge_environment_settings = requests.Session.merge_environment_settings
-
-    def merge_environment_settings(self, url, proxies, stream, verify, cert):  # pylint: disable=R0913
-        # Verification happens only once per connection so we need to close
-        # all the opened adapters once we're done. Otherwise, the effects of
-        # verify=False persist beyond the end of this context manager.
-        opened_adapters.add(self.get_adapter(url))
-
-        settings = old_merge_environment_settings(self, url, proxies, stream, verify, cert)
-        settings["verify"] = False
-
-        return settings
-
-    requests.Session.merge_environment_settings = merge_environment_settings
-
-    try:
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", InsecureRequestWarning)
-            yield
-    finally:
-        requests.Session.merge_environment_settings = old_merge_environment_settings
-
-        for adapter in opened_adapters:
-            try:
-                adapter.close()
-            except:  # pylint: disable=W0702
-                pass
